fix: change purchase limit type to uint32

+++ contracts/Option/IOddzAsset.sol
@@ -11,7 +11,7 @@ interface IOddzAsset {
     event NewAssetPair(uint32 indexed _id, uint32 indexed _primary, uint32 indexed _strike, bool _status);
     event AssetActivatePair(uint32 indexed _id, uint32 indexed _primary, uint32 indexed _strike);
     event AssetDeactivatePair(uint32 indexed _id, uint32 indexed _primary, uint32 indexed _strike);
-    event SetPurchaseLimit(uint32 indexed _id, uint32 indexed _primary, uint32 indexed _strike, uint256 _limit);
+    event SetPurchaseLimit(uint32 indexed _id, uint32 indexed _primary, uint32 indexed _strike, uint32 _limit);
 
     struct Asset {
         uint32 _id;
@@ -26,7 +26,7 @@ interface IOddzAsset {
         uint32 _primary;
         uint32 _strike;
         bool _active;
-        uint256 _limit;
+        uint32 _limit;
     }
 
     /**
@@ -65,7 +65,7 @@ interface IOddzAsset {
      * @param _limit Purchase limit for the pair
      * @return pairId Asset pair ID
      */
-    function addAssetPair(uint32 _primary, uint32 _strike, uint256 _limit) external returns (uint32 pairId);
+    function addAssetPair(uint32 _primary, uint32 _strike, uint32 _limit) external returns (uint32 pairId);
 
     /**
      * @notice Activate an asset pair
@@ -88,5 +88,5 @@ interface IOddzAsset {
      * @param _pairId ID of an valid asset pair
      * @param _limit Purchase limit for an asset pair
     */
-    function setPurchaseLimit(uint32 _pairId, uint256 _limit) external; 
+    function setPurchaseLimit(uint32 _pairId, uint32 _limit) external; 
 }
